+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # Unreleased
 
-* Fix interactive carets in rich texts.
-  - Now transfers pointer capture to the caret widget of each leaf text as they become the active focus. !!: TODO
+* Fix interactive carets in rich texts losing pointer capture when crossing over leaf texts.
 * Fix interactive insert caret appearing in non editable text.
-* Update `webrender` and  `swgl`.
+* Update renderer dependencies.
 
 # 0.15.3